requirements.txt
flask==2.0.1
pymongo==4.0.1
python-dotenv==0.19.0
scikit-learn==0.24.2
numpy==1.21.2 